fixed actioncable errors

+++ app/controllers/purchases_controller.rb
@@ -8,20 +8,29 @@ class PurchasesController < ApplicationController
   def new
   end
 
-  def create    
-    #binding.pry
-    current_user.purchases.where(room_id: @room.id, invoice_id: nil).destroy_all
-    items_number = @items_number.to_a
-    variable_params.each do |index, nitems|
-      @purchase.items << Item.where(product_id: items_number[index.to_i][0], sold: false, room_id: @room.id).limit(nitems.to_i)
-    end 
-    if @purchase.save
-      flash[:notice] = "Ihr Kaufangebot wurde gespeichert! Jetzt musst du auf den Verkäufer warten"
-      redirect_to room_path(@room)
-    else
-      flash[:error] = "Ein Fehler ist aufgetreten, der Kauf wurde nicht gespeichert"
-      redirect_to room_path(@room)
-    end           
+  def create  
+    # perform validation on price, not more then available resources  
+    unless current_user.validation_balance(@price)
+      flash[:error] = "Der Kaufbetrag ist zu hoch! Du müsst weniger Ressources nutzten"
+      redirect_to room_path(@room)      
+    else 
+      @price.save
+
+      # create purchase
+      current_user.purchases.where(room_id: @room.id, invoice_id: nil).destroy_all
+      @purchase = Purchase.new(room_id: params[:room_id], price_id: @price.id, user_id: current_user.id)
+      items_number = @items_number.to_a  
+      variable_params.each do |index, nitems|
+        @purchase.items << Item.where(product_id: items_number[index.to_i][0], sold: false, room_id: @room.id).limit(nitems.to_i)
+      end 
+      if @purchase.save
+        flash[:notice] = "Ihr Kaufangebot wurde gespeichert! Jetzt musst du auf den Verkäufer warten"
+        redirect_to room_path(@room)
+      else
+        flash[:error] = "Ein Fehler ist aufgetreten, der Kauf wurde nicht gespeichert"
+        redirect_to room_path(@room)
+      end     
+    end      
   end
 
   def edit
@@ -64,18 +73,7 @@ class PurchasesController < ApplicationController
   def set_variables
     @room = Room.find(params[:room_id])    
     @price = Price.new(price_params)
-
-    # you can not do this because price is an object.. maybe you
-    # can create a method in the price.rb model ?
-    unless current_user.validation_balance(@price)
-      flash[:error] = "Der Kaufbetrag ist zu hoch! Du müsst weniger Ressources nutzten"
-      redirect_to room_path(@room)      
-    else 
-      @price.save
-    end
-
-    @items = @room.items.where(sold: false, used: false).order(:product_id) #.where("room_id != :nil OR bonus = :true", {nil: nil, true: true})
-    @purchase = Purchase.new(room_id: params[:room_id], price_id: @price.id, user_id: current_user.id)
-    @items_number = @items.group(:product_id).count
+    @items = @room.items.where(sold: false, used: false).order(:product_id)
+    @items_number = @items.group(:product_id).count          
   end    
 end
